Refactoring: move formatting to new file
